(SC-7) Moved the <grid>, renamed to "model.ts"
