+++ src/app/App.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
-import {Grid} from "@material-ui/core";
 import {ChampionPage} from "./components/champion";
 import {Header} from "./components/commons/Header";
 
@@ -10,11 +9,7 @@ export const App = () => {
             <Header/>
             <Switch>
                 <Route path="/champion">
-                    <Grid container spacing={3}>
-                        <Grid item xs={2}>
-                            <ChampionPage/>
-                        </Grid>
-                    </Grid>
+                    <ChampionPage/>
                 </Route>
                 <Route path="/">
                 </Route>
